// shared/types.ts
export interface SaleItem {
  productId: string;
  productName: string;
  quantity: number;
  returnableQuantity: number; // Quantidade que ainda pode ser devolvida
  unitPrice: number;
  totalPrice: number;
  imageUrl?: string;
}

// Representa uma venda completa
export interface Sale {
  id:string;
  date: string; // ISO string format
  items: SaleItem[];
  totalAmount: number;
  paymentMethod: 'Dinheiro' | 'Cartão de Crédito' | 'Cartão de Débito' | 'PIX' | 'A Pagar na Entrega' | 'Troca / Vale-Crédito';
  customerId?: string;
  customerName?: string;
  deliveryId?: string; // ID da entrega associada
  status: 'Completed' | 'Partially Returned' | 'Fully Returned' | 'Pending Payment';
  storeCreditAmountUsed?: number; // Valor pago com vale-crédito
  // Dados da NFC-e
  nfceAccessKey?: string;
  nfceQrCodeUrl?: string;
}

// Representa um item devolvido
export interface ReturnItem {
  productId: string;
  productName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}
// ...
export interface User {
  id: string;
  name: string;
  email: string;
  role: 'administrador' | 'vendedor' | 'caixa';
  password?: string; // Adicionado para a simulação de login
}